pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "opticiq"
dynamic = ["version"]
description = "Optical Image Quality and Beam Tests routines"
license = {text = "BSD-2-Clause"}
authors = [
    {name = "Chris Cannon", email = "chris.cannon.9001@gmail.com"}
]
readme = "README.rst"
requires-python = ">= 3.7"
# from reportlab.graphics.shapes import Polygon doesn't seem to work in py 3.7
dependencies = [
    'numpy',
    'matplotlib',
    'PIllow',
    'scikit-image',
    'sympy',
    'more-itertools;python_version<"3.10"' # itertools.pairwise in python>=3.10
]

[tool.setuptools.dynamic]
version = {attr = "opticiq.__version__"}

[project.optional-dependencies]
test = [
    "pytest",
]
gen = [
    "gdstk",
    "reportlab"
]

[project.urls]
Homepage = "https://github.com/chriscannon9001/opticiq"
